move ozoneClientImpl to a separate file
